Move Image Embeding into Extension
